Created HomePage and ProfilePage
+ moved Login to separate page
