+++ I challenge you 3.0/MainWindow.xaml
@@ -11,70 +11,7 @@
         Width="1000"
         ResizeMode="CanMinimize">
     <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-        <Label 
-            Content="Login"
-            HorizontalAlignment="Center" 
-            Margin="0,51,0,0"
-            VerticalAlignment="Top" 
-            FontWeight="Bold"
-            FontSize="30"
-            Grid.Row="0"
-            />
-
-        <Label 
-            Content="Username:"
-            HorizontalAlignment="Left" 
-            Margin="301,38.2,0,0"
-            VerticalAlignment="Top" 
-            FontWeight="Bold"
-            FontSize="18"
-            Grid.Row="1"
-            />
-
-        <TextBox x:Name="username"
-            HorizontalAlignment="Center" 
-            TextWrapping="Wrap" 
-            VerticalAlignment="Top"
-            Width="164"
-            Margin="321,41,276.6,0" 
-            Height="31"
-            FontSize="18" 
-            Grid.Row="1"
-            />
-        <Label 
-            Content="Password:"
-            HorizontalAlignment="Left" 
-            Margin="301,91.2,0,0"
-            VerticalAlignment="Top" 
-            FontWeight="Bold"
-            FontSize="18" RenderTransformOrigin="1.552,0.477"
-            Grid.Row="1"
-            />
-        <PasswordBox
-            Name="password"
-            HorizontalAlignment="Center"
-            Margin="321,91,276.6,0"
-            VerticalAlignment="Top"
-            Width="164"
-            Height="31"
-            Grid.Row="1"/>
-        <Button
-            x:Name="Login"
-            Content="Login" 
-            HorizontalAlignment="Center"
-            VerticalAlignment="Top" 
-            Margin="452,35.2,441.6,0"
-            Height="35"
-            Width="100" 
-            FontWeight="Bold"
-            FontSize="18" 
-            Click="Login_Click"
-            Grid.Row="2"
-           />
+        <Frame NavigationUIVisibility="Hidden" x:Name="frame" Source="Pages/LoginPage.xaml">
+        </Frame>
     </Grid>
 </Window>
