+++ I challenge you 3.0/MainWindow.xaml.cs
@@ -1,6 +1,4 @@
-﻿using I_challenge_you_3._0.DataAccessLayers;
-using I_challenge_you_3._0.Model;
-using System.Windows;
+﻿using System.Windows;
 
 namespace I_challenge_you_3._0
 {
@@ -12,19 +10,6 @@ namespace I_challenge_you_3._0
         public MainWindow()
         {
             InitializeComponent();
-        }
-
-        private void Login_Click(object sender, RoutedEventArgs e)
-        {
-            User user = UserDAL.getUser(username.Text, password.Password);
-            if (user == null)
-            {
-                MessageBox.Show("The username or password you entered are wrong, please check them and try again");
-            }
-            else
-            {
-                MessageBox.Show(user.Email + " " + user.Username);
-            }
         }
     }
 }
