+++ I challenge you 3.0/Pages/HomePage.xaml
@@ -1,0 +1,21 @@
+﻿<Page x:Class="I_challenge_you_3._0.HomePage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:I_challenge_you_3._0"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="HomePage">
+
+    <Grid>
+        <Button
+            Height="100"
+            Width="100"
+            VerticalAlignment="Top"
+            HorizontalAlignment="Right"
+            Click="Profile_Click"
+            Content="{Binding loggedUser.Username}">
+        </Button>
+    </Grid>
+</Page>
