+++ I challenge you 3.0/Pages/LoginPage.xaml.cs
@@ -1,0 +1,44 @@
+﻿using I_challenge_you_3._0.DataAccessLayers;
+using I_challenge_you_3._0.Model;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace I_challenge_you_3._0
+{
+    /// <summary>
+    /// Interaction logic for LoginPage.xaml
+    /// </summary>
+    public partial class LoginPage : Page
+    {
+        public LoginPage()
+        {
+            InitializeComponent();
+        }
+
+        private void Login_Click(object sender, RoutedEventArgs e)
+        {
+            User user = UserDAL.getUser(username.Text, password.Password);
+            if (user == null)
+            {
+                MessageBox.Show("The username or password you entered are wrong, please check them and try again");
+            }
+            else
+            {
+                MessageBox.Show(user.Email + " " + user.Username);
+                NavigationService.Navigate(new HomePage(user));
+            }
+        }
+    }
+}
