+++ I challenge you 3.0/Pages/ProfilePage.xaml
@@ -1,0 +1,25 @@
+﻿<Page x:Class="I_challenge_you_3._0.Pages.ProfilePage"
+      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+      xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+      xmlns:local="clr-namespace:I_challenge_you_3._0.Pages"
+      mc:Ignorable="d" 
+      d:DesignHeight="450" d:DesignWidth="800"
+      Title="ProfilePage">
+
+    <Grid>
+        <Label 
+            Height="32" 
+            Width="360"
+            Margin="0 0 0 100"
+            Content="{Binding loggedUser.Username}">
+        </Label>
+        <Label 
+            Height="32"
+            Width="360"
+            Margin="0 0 0 0"
+            Content="{Binding loggedUser.Email}">
+        </Label>
+    </Grid>
+</Page>
