+++ I challenge you 3.0/bin/Debug/I challenge you 3.0.exe.config
@@ -4,7 +4,7 @@
         <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2" />
     </startup>
 	<connectionStrings>
-		<add connectionString="Data Source=DESKTOP-AUUCNDM;Initial Catalog=ChallengeYouDatabase;Integrated Security=SSPI;"
+		<add connectionString="Data Source=LAPTOP-MARIUS;Initial Catalog=ChallengeYouDatabase;Integrated Security=SSPI;"
 			 name="myConStr"/>
 	</connectionStrings>
 </configuration>
